app/Back End/requirements.txt
pandas
matplotlib
seaborn
numpy
flask
requests
python-dotenv
render_template
Tkinter